{"schema":"decodex/commit/1","summary":"Split radar validation and ledger modules","authority":"manual"}
